email refactor
